+++ src/main/java/org/truffleruby/extra/TruffleRubyNodes.java
@@ -9,23 +9,18 @@
  */
 package org.truffleruby.extra;
 
-import org.jcodings.specific.USASCIIEncoding;
 import org.truffleruby.RubyContext;
 import org.truffleruby.RubyLanguage;
 import org.truffleruby.builtins.CoreMethod;
 import org.truffleruby.builtins.CoreMethodArrayArgumentsNode;
 import org.truffleruby.builtins.CoreMethodNode;
 import org.truffleruby.builtins.CoreModule;
-import org.truffleruby.core.rope.RopeOperations;
-import org.truffleruby.core.string.StringOperations;
 import org.truffleruby.extra.ffi.Pointer;
-import org.truffleruby.shared.BuildInformationImpl;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.TruffleOptions;
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.object.DynamicObject;
 
 @CoreModule("TruffleRuby")
 public abstract class TruffleRubyNodes {
@@ -62,20 +57,6 @@ public abstract class TruffleRubyNodes {
         @TruffleBoundary
         public static boolean isSulongAvailable(RubyContext context) {
             return context.getEnv().isMimeTypeSupported(RubyLanguage.LLVM_BITCODE_MIME_TYPE);
-        }
-
-    }
-
-    @CoreMethod(names = "revision", onSingleton = true)
-    public abstract static class RevisionNode extends CoreMethodArrayArgumentsNode {
-
-        @TruffleBoundary
-        @Specialization
-        protected DynamicObject revision() {
-            return StringOperations.createFrozenString(
-                    getContext(),
-                    RopeOperations
-                            .encodeAscii(BuildInformationImpl.INSTANCE.getShortRevision(), USASCIIEncoding.INSTANCE));
         }
 
     }